requirements.txt
# ============================================================
# 3DGS Digital Twin — Python Dependencies
# ============================================================

# Core numerical / vision
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
plyfile>=0.9

# 3D geometry
open3d>=0.17

# Clustering / ML
scikit-learn>=1.3

# Pipeline config
PyYAML>=6.0

# Progress bars (used in DBSCAN steps)
tqdm>=4.66

# --- Segmentation (Step 1) ---
# Install SAM from source:
#   pip install git+https://github.com/facebookresearch/segment-anything.git
# Then download weights:
#   wget https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth
torch>=2.0          # required by SAM
torchvision>=0.15   # required by SAM
# segment-anything   # install manually (see above)

# --- ROS 2 workspace (ros2_ws/) ---
# Managed separately via colcon; see ros2_ws/mtc_controller_final/src/README.md
